Remove the `assert` package.
