--- assert/doc.go
@@ -1,2 +1,0 @@
-// Package assert contains the assertions used by the test runner.
-package assert
